extract PTRSolver abstract base, rename CVXPy impl, add backend discriminator
Prep for a second PTR backend (QPAX) by separating the PTR formulation from
the CVXPy assembly:

- `openscvx/solvers/ptr_solver.py` is now an ABC owning `PTRSolveResult` and
  the per-iteration update_* / solve contract, plus a `_scaling(unified)`
  staticmethod hoisted out of the old `create_variables`.
- `openscvx/solvers/cvxpy_ptr_solver.py` (new) houses `CVXPyPTRSolver` —
  a verbatim move of the previous `PTRSolver` impl, minus the obsolete
  "Future Backend Support" docstring note.
- `SolverSpec` becomes `PTRSolverSpec` with a `backend: Literal["cvxpy",
  "qpax"]` discriminator; a model validator rejects CVXPy-only fields when
  `backend="qpax"`. `SolverSpec` kept as a deprecated `__getattr__` alias.
- `_SOLVER_MAP` and the unused `ConvexSolver.cvx_solver` annotation are
  removed.
- `printing.py` reads the solver class name and `cvx_solver` via `getattr`
  so the backend display still works for QPAX (no `cvx_solver` field).
- `problem.py` / `loader.py` updated for the new spec shape; `ox.PTRSolver`
  and `ox.CVXPyPTRSolver` exported, `ox.QPAXPTRSolver` lazy via
  `__getattr__`.

No behavior change on the CVXPy path — tests/test_brachistochrone.py and
tests/test_loader.py both pass unchanged.
