+++ openscvx/solvers/base.py
@@ -6,33 +6,21 @@ must follow for use within successive convexification algorithms.
 !!! note
 
     Solvers own their optimization variables via ``create_variables()``.
-    Convex constraint lowering remains in ``lower.py`` but uses the solver's
-    variables.
+    Convex constraint lowering remains in ``openscvx.symbolic.lower`` but uses
+    the solver's variables — non-CVXPy backends that don't support user
+    ``.convex()`` constraints (e.g.
+    :class:`openscvx.solvers.qpax_ptr_solver.QPAXPTRSolver`) raise in
+    ``initialize()`` instead of going through a lowerer.
 
-    When adding non-CVXPy backends, there are two approaches:
-
-    1. **Solver owns the lowerer**: The solver implements a
-       ``lower_convex_constraints()`` method containing the lowering logic.
-
-    2. **Solver determines the lowerer**: The solver references which lowerer
-       to use, but the lowering logic stays in ``lower.py``. Example:
-
-       ```python
-       # In solver
-       @property
-       def lowerer(self):
-           from openscvx.symbolic.lower import lower_cvxpy_constraints
-           return lower_cvxpy_constraints
-
-       # In lower_symbolic_problem()
-       lowered_constraints = solver.lowerer(constraints, solver.variables, parameters)
-       ```
+    See :class:`openscvx.solvers.ptr_solver.PTRSolver` for the PTR-specific
+    interface every PTR backend implements.
 """
 
+import warnings
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any, Dict, List, Literal, Optional
 
-from pydantic import BaseModel, ConfigDict
+from pydantic import BaseModel, ConfigDict, model_validator
 
 if TYPE_CHECKING:
     from openscvx.config import Config
@@ -90,10 +78,6 @@ class ConvexSolver(ABC):
                     self._prob.solve()
                     return MyResult(...)
     """
-
-    #: Backend solver name (e.g., ``"QOCO"``, ``"CLARABEL"``).  Subclasses
-    #: must set this in ``__init__``.
-    cvx_solver: str
 
     @abstractmethod
     def create_variables(
@@ -241,11 +225,17 @@ class ConvexSolver(ABC):
 # Pydantic spec for dict / YAML validation
 # =============================================================================
 
-_SOLVER_MAP: Dict[str, type] = {}  # populated by __init__.py after all classes are imported
 
+class PTRSolverSpec(BaseModel):
+    """Validates PTR solver configuration from dict/YAML input.
 
-class SolverSpec(BaseModel):
-    """Validates solver configuration from dict/YAML input.
+    The ``backend`` discriminator selects which concrete PTR backend to build:
+    ``"cvxpy"`` (the default,
+    :class:`openscvx.solvers.cvxpy_ptr_solver.CVXPyPTRSolver`) or ``"qpax"``
+    (:class:`openscvx.solvers.qpax_ptr_solver.QPAXPTRSolver`).
+
+    ``cvx_solver``, ``cvxpygen``, and ``cvxpygen_override`` are CVXPy-only;
+    setting them under ``backend="qpax"`` is a configuration error.
 
     !!! warning
         Enabling ``cvxpygen`` currently disables sparse parameter declarations.
@@ -255,15 +245,61 @@ class SolverSpec(BaseModel):
     """
 
     type: Literal["PTRSolver"] = "PTRSolver"
-    cvx_solver: str = "QOCO"
+    backend: Literal["cvxpy", "qpax"] = "cvxpy"
+    cvx_solver: Optional[str] = None
     solver_args: Optional[Dict[str, Any]] = None
     cvxpygen: bool = False
     cvxpygen_override: bool = False
 
     model_config = ConfigDict(extra="forbid")
 
+    @model_validator(mode="after")
+    def _check_backend_fields(self):
+        if self.backend == "qpax":
+            offenders = [
+                name
+                for name, value in (
+                    ("cvx_solver", self.cvx_solver),
+                    ("cvxpygen", self.cvxpygen),
+                    ("cvxpygen_override", self.cvxpygen_override),
+                )
+                if value
+            ]
+            if offenders:
+                raise ValueError(
+                    f"{offenders} only valid for backend='cvxpy'; "
+                    "remove these fields or set backend='cvxpy'."
+                )
+        return self
+
     def build(self) -> ConvexSolver:
-        cls = _SOLVER_MAP.get(self.type)
-        if cls is None:
-            raise ValueError(f"Unknown solver {self.type!r}; expected one of {sorted(_SOLVER_MAP)}")
-        return cls(**self.model_dump(exclude={"type"}, exclude_unset=True))
+        # Local imports keep CVXPy / qpax out of the import path until the
+        # corresponding backend is actually requested.
+        if self.backend == "cvxpy":
+            from .cvxpy_ptr_solver import CVXPyPTRSolver
+
+            return CVXPyPTRSolver(
+                cvx_solver=self.cvx_solver or "QOCO",
+                solver_args=self.solver_args,
+                cvxpygen=self.cvxpygen,
+                cvxpygen_override=self.cvxpygen_override,
+            )
+        from .qpax_ptr_solver import QPAXPTRSolver
+
+        return QPAXPTRSolver(solver_args=self.solver_args)
+
+
+def __getattr__(name: str):
+    """Deprecated alias: ``SolverSpec`` → :class:`PTRSolverSpec`.
+
+    Kept for one release so existing dict/YAML configs and tests that import
+    ``SolverSpec`` continue to work. Emit a ``DeprecationWarning`` on access.
+    """
+    if name == "SolverSpec":
+        warnings.warn(
+            "openscvx.solvers.base.SolverSpec is deprecated; use PTRSolverSpec.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return PTRSolverSpec
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
